hotfix: modularização das funções do robô
