+++ src/firmware/NDCbot/modules/__init__.py
@@ -1,0 +1,7 @@
+# Constantes compartilhadas
+LOG_SUCESSO = 0
+LOG_QR_NAO_CORRESPONDE = 1
+LOG_QR_NAO_RECONHECIDO = 2
+LOG_REMEDIO_NAO_COLETADO = 3
+LOG_REMEDIO_CAIU = 4
+LOG_ERRO_DOBOT = 5
